encapsulated full_body module + new stances
Creating single full_body module that accepts parameters
for the position of all joints to make it easier to create
new poses.

Base now is part of full_body and connects to a specific
foot instead of positioning relative to the body.

Also adding several new poses.

Note: This is still catching up on changes made previously
but not committed.